Proxy Update: Contexts
We are now passing request contexts down the chain to the invoker, so that middlewares can use the values. Rather than creating a custom map, we are going to assume that the application is providing the information it wants us to have with http middlewares. This is in line with standard Go web application development.

Requests are now tied to ProxyRequest, which will keep the appropriate version of the config for the life of the request. Request/Response handling is now the responsibility of the ProxyRequest. The Proxy is merely providing initialization data to the ProxyRequest, and reading it's final responses to send to back to the client.

Fasthttp has been moved to hack, as it is not a priority for development at the moment.