Update Document.PrintValue to print block strings using triple quotes
Fixes #292.

GraphQL has two types of string literals: double-quoted strings and
triple-quoted "block" strings. Block strings can include unescaped white
space, lint terminators, quote and backslash characters. (These
characters must be escaped in double-quoted strings). See the String
Value section of the GraphQL spec[1] for details.

[1] https://spec.graphql.org/June2018/#sec-String-Value)

This PR updates the Document.PrintValue function so it emits a
triple-quoted string for block strings in the AST. That way printed
block strings, which may contain unescape characters, will always be
valid. This assumes the string content was valid when the AST was
constructed (which seems like a reasonable assumption).

While working on this PR I realized the Document.ValueToJSON method also
has a bug. Raw string content in the AST is unquoted in block strings,
but the ValueToJSON code assumes string content is quoted. This means
invalid JSON may be produced when block strings are serialized to JSON.
This bug IS NOT addressed in this PR. Instead, I'll create a separate
issue to address this bug.