Add additional LocalTypeFieldExtractor tests and benchmark
This is a preliminary PR for #293 and #294. The plan is to fix those
issues by rewriting the LocalTypeFieldExtractor.GetAllNodes method. This
PR adds tests to nail down the current behavior. While writing the tests
I noticed a few unhandled edge cases in addition to issue #293: concrete
interface types and union members aren't included in node children if
the types are only reachable from a root node via an interface or union.
It looks like these edge cases were partially addressed as part of #283.
The rewrite will address them as well. The union tests are currently
skipped because they don't pass at all. The issue #293 tests are also
skipped, for the same reason.

The PR also updates the existing tests so root and child nodes are
sorted before comparison. This is because the updated algorithm will
very likely traverse nodes and fields in a different order and it seems
like it would be good to not change the tests too much in the same diff
that also modifies GetAllNodes.

Finally, the current benchmark results for the GetAllNodes method are as
follows:
```
$ cd pkg/engine/plan
$ go test -bench=. -run=xxxx
goos: darwin
goarch: amd64
pkg: github.com/jensneuse/graphql-go-tools/pkg/engine/plan
BenchmarkGetAllNodes-8  111  11224753 ns/op  24116372 B/op  10363 allocs/op
PASS
ok  github.com/jensneuse/graphql-go-tools/pkg/engine/plan  2.463s
```