Include `@requires` fields when constructing entity representations
This PR fixes #314.

graphql-go-tools currently doesn't include `@requires` fields in entity
representations. This PR updates the GraphQL datasource planner to
include `@requires` fields when constructing representations for an
upstream operation.

The representations logic has been updated to use RequiresFields from
plan.FieldConfiguration. This field already includes the necessary
federation keys and required fields.

Due to the order the Planner visitor calls LeaveDocument, the
representations are updated each time a field is visited instead of
being set once during LeaveDocument. This is because ConfigureFetch is
called before LeaveDocument. It would be possible to move the code to
LeaveDocument if the visit order was reversed for the Leave* methods,
but that is a large change so it wasn't made in this PR.

Note that the use of RequiresFields also fixes a bug where
graphql-go-tools didn't align with the federation spec with respect to
federation keys on extended types. The existing representations code
includes all the keys defined on a type, but the spec states that
exactly one key must match in a subgraph[1]. The RequiresFields logic
aligns with this requirement. The GraphQL datasource tests have been
updated accordingly.

[1] https://www.apollographql.com/docs/federation/entities/#referencing