Improve variable renderer JSON validation messaging
This PR fixes #323. It updates the jsonschema.Validate function to
return an error when validation fails (instead of a boolean). This error
contains details about why the validation failed.

For example, previously the error surfaced in the GraphQL response
looked like:
```
{
    "errors": [{
        "message": "json schema validation failed on Variable Renderer"
    }]
}
```

Now it looks like:
```
{
    "errors": [{
        "message": "could not render JSON variable, validation failed: /exerciseId: type should be string, got null",
    }]
}
```

(For a JSON object with a top-level key "exerciseId" that is a required
string.)